Fix a "Source Unauthorized" error that happens with the Microsoft iSNS server
When two targets are registered with isns_target_register(), the first
target is correctly registered but the registration of the second
target fails. The Microsoft iSNS server returns the following error:
"Source Unauthorized".

The Microsoft iSNS server requires that the same "source attribute"
(iqn) is used to register both targets.

See the link below for details:

  http://www.ietf.org/mail-archive/web/ips/current/msg00997.html

Signed-off-by: Christophe Vu-Brugier <cvubrugier@yahoo.fr>