Remove use_isns and use_isns_ac variables
The use_isns variable is useless because we are an iSNS client :-)

Moreover, I also remove use_isns_ac because I don't know how to
handle it correctly.

Setting it to true allows to delegate access control to the iSNS
server. My (limited) understanding is that iSNS access control
requires LIO to ask the iSNS server before granting access to a
target. It probably cannot be done easily with the LIO iSCSI target
running in kernel space and an iSNS client running in userspace.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@yahoo.fr>