Fix the allocation of epoll events
Valgrind reports that some data are uninitialized:

  1 errors in context 1 of 4:
  Syscall param epoll_ctl(event) points to uninitialised byte(s)
     at 0x4F17FBA: epoll_ctl (syscall-template.S:81)
     by 0x4067C4: epoll_add_fd (target-isns.c:71)
     by 0x406ABF: main (target-isns.c:180)
   Address 0xffeffff48 is on thread 1's stack
   Uninitialised value was created by a stack allocation
     at 0x40671F: epoll_add_fd (target-isns.c:52)

The epoll_event structure that valgrind reports as not being fully
initialized is allocated on the stack. This is wrong because the
structure can be accessed by epoll after the function returns. It is
safer to declare a static array of epoll_event structures.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@yahoo.fr>