Fix erroneous handling of iSNS socket descriptors
In the code, the value 0 is used to reset the socket file descriptors
and has the semantic of "no file descriptor". However, since 0 is a
valid file descriptor, this usage causes some subtle bug like issue #8
where syslog messages are written to the iSNS socket.

This patch updates the code to use -1 to mean "no file descriptor".

Signed-off-by: Christophe Vu-Brugier <cvubrugier@yahoo.fr>