Maintain a single list of portals
This commit introduces the "tpg_portal" struct which is an associative
entity between a TPG and a portal. This entity allows to implement a
"many to many" association between TPGs and portals: a TPG can be
associated with several portals and several TPGs can be associated
with a single portal.

Also this commit drops the "portals cache" that was implemented to
ensure a portal is registered only once in a DevAttrReg request.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>