Mention that tarballs are available on Fedora Hosted in README.md
GitHub is really convenient for collaboration and source code. But it
is not really good at making releases.

With GitHub, a release tarball is either automatically generated or
uploaded by the maintainer. The problem with an "automatically
generated" tarball is that its hash is not stable because the content
of the archive may change (due to file ordering). The problem with an
"uploaded" tarball on GitHub is that there is no easy way to put the
hash file close to the tarball.

As a consequence, the future releases of target-isns will be made
available on Fedora Hosted.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>