Set C_STANDARD target property only for CMake 3.1 or newer
The C_STANDARD target property was added in CMake 3.1. The compilation
using older versions of CMake, like 2.8.12 on Ubuntu 14.04 fails because
the -std=c99 C flag is not added. To fix the problem, for CMake older
than 3.1, add the -std=c99 to CMAKE_C_FLAGS, for newer versions set
C_STANDARD target property.

Signed-off-by: Victor Dodon <dodonvictor@gmail.com>