Upgrade version to 0.6.3
The version number is bumped from 0.6.1 to 0.6.3 because the
intermediate version 0.6.2 already exists as a tag (v0.6.2). However,
the version number was only updated to 0.6.2 in the v0.6.2 tag which
explains why the master branch version number is lagging behind.

v0.6.2 was tagged on Friday Oct 16 2015 and there was not much activity
on target-isns since then. The changes were mostly cosmetic changes.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>