Refactor the data structures of TPGs and portals
This patch replaces the "tpg_portal" structure which was meant as an
associative entity between a TPG and a portal with a more logical data
structure hierarchy. A TPG now contains a list of pointers to
portals. A global list of portals is still maintained and several TPGs
can contain the same portal by pointing to it. With that, it becomes
much easier to iterate over the portals of a TPG.

Moreover, a portal now has a reference counter which denotes how many
TPGs point to it. This will be used in a future commit.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>