A portal is no longer registered when its refcount reaches zero
According to RFC 4171 (iSNS)

  5.6.5.4. Device Deregister Request (DevDereg)

  Upon receiving the DevDereg, the iSNS server removes all objects
  identified by the Operating Attribute(s), and all subordinate
  objects that are solely dependent on those identified objects.

The iSNS server removes a portal when no target depends on it. Thus,
we have to mark a portal as no longer registered when its reference
counter reaches zero so that when a new target is later registered,
the portal is registered too.

This patch fixes issue #38.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>