Set the "replace" flag when updating the registration of a target
Keep track of which targets are registered and use this information
when updating the registration of a target to set the iSNS replace
bit. With that, the Microsoft iSNS server does not return an invalid
update error (status 14) and portals that no longer exist are
correctly removed.

This patch fixes issue #40.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>