Fix portal registration one more time
With the addition of the iSNS replace bit to update the registration
of a target, Open-iSNS may complain that a TPG depends on an unknown
portal.

Consider the following scenario:

1. create a target (default portal) -> registration OK
2. delete the default portal -> registration OK (replace bit)
3. create portal 1.2.3.4:3260 -> registration OK (replace bit)
4. create portal 5.6.7.8:3260 -> registration FAIL (replace bit)

The problem at step 4 is that the portal 1.2.3.4:3260 created during
step 3 must be registered again because the iSNS replace bit forces
the iSNS server to deregister it.

Moreover, portals cannot be unconditionally registered because the
Microsoft iSNS server expects that "an object specified in a
DevAttrReg request to update an Entity [should not exist] in another
Entity." [1]

As a consequence, this patch removes the "registered" and "refcount"
fields from "struct portal" and uses another condition: check that the
portal is not used by another target which is already registered.

[1] https://docs.microsoft.com/en-us/previous-versions/windows/hardware/design/dn653564(v=vs.85)#invalid-update-status-code-14

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>