Register the iSCSI alias attribute
The target alias is a user-friendly string associated with the iSCSI
qualified name.

In LIO, the target alias is not a parameter of the target (there is no
target parameter), but a parameter of the TPG. Because several TPGs
may be associated with a given target, conflicting "TargetAlias"
parameters may exist. The logic implemented in this patch is to select
the first target alias encountered which is different from the default
target alias in LIO ("LIO Target"). If no custom target alias is
found, the default target alias is selected.

This patch fixes issue #48.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>