Fix connection with IPv6 servers
If an IPv6 address is specified, it is truncated when the IP address
information is copied to the global "ss" variable. When that variable
is later read to connect() to the server, the IP address it contains
is wrong and the connection fails.

This is fixed by using the "ai_addrlen" field to determine how much
bytes from "ai_addr" should be copied to the "sockaddr_storage"
structure.

Part of issue #51.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>